# ApiAulearning/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aulearning API",
        "description": "API de Aulearning",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Listar calificaciones",
                "operationId": "219f2a64ee0e5767a4aa4f1c65debdbc",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "course_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_grade",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_grade",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortDirection"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de calificaciones"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Crear calificación",
                "operationId": "a41924c58474d7924b256fe28834ba77",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grade",
                                    "student_id",
                                    "course_id"
                                ],
                                "properties": {
                                    "grade": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.5
                                    },
                                    "student_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Calificación creada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/grades/{id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Ver calificación",
                "operationId": "a551c7d5536a126502162cd79d1b462c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calificación encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Grades"
                ],
                "summary": "Actualizar calificación",
                "operationId": "593caae352e2a275a2d4a8e632038ccf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calificación actualizada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Eliminar calificación",
                "operationId": "0209e0f4a4c0932f4e118311a6103245",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calificación eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Listar notificaciones",
                "operationId": "b1274a2aaae8a4295a7e1925e8d8090f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "read",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortDirection"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de notificaciones"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Crear notificación",
                "operationId": "0faafeb87df4ff1447878720cce80571",
                "responses": {
                    "201": {
                        "description": "Notificación creada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Ver notificación",
                "operationId": "6293e7ffc395913cd17da6467e486699",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Eliminar notificación",
                "operationId": "0f28a0400610f81f01e2d25ce4e0edff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marcar como leída",
                "operationId": "356012fe06901053b88e36ec633c6524",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación leída"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Número de elementos por página",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 15,
                    "minimum": 1
                }
            },
            "SortBy": {
                "name": "sort_by",
                "in": "query",
                "description": "Campo por el que ordenar",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "id"
                }
            },
            "SortDirection": {
                "name": "sort_direction",
                "in": "query",
                "description": "Dirección de ordenación",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "desc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}